calcphase v1.0
Calculate phase of spectrum based on the first complex point of the
FID. This is primarily designed for cases in which the first point
of the FID occurs exactly at time zero, i.e., an echo. In this case,
calcphase may produces answers which are more reliable than aph
or aph0.