lcdatast v1.0
Expands the functionality of the "lcdatast" macro to properly adjust
the UV chromatogram on the SUN screen for expansions of on-flow or
stop-flow measurements (with the standard macro, the UV is always
displayed for the whole time of the chromatography). The contribution
also makes "lcdatast" label intensities in the UV chromatogram
correctly.